+++ docs/changelog.rst
@@ -35,6 +35,8 @@ Minor changes
 
 * Remove ``cpt.Nemoh()`` class that was replaced by :class:`~capytaine.bem.solver.BEMSolver` in version 1.1 (:pull:`291`)
 
+* Add function :func:`~capytaine.bem.airy_waves.airy_waves_free_surface_elevation` to compute the free surface elevation at points (:pull:`293`).
+
 -------------------------------
 New in version 1.5 (2022-12-13)
 -------------------------------